added sample config
(cherry picked from commit 6b49b802dd762efc19e0e86e40a91c8d11cc5277)
